파일 업로드, 저장 및 삭제
1. 클라이언트에 파일 전송하는 방법 2가지 - send(), 스트리밍
2. 서버에 파일 저장하는 방법 - pdfkit이란 npm 사용
3. 품목 이미지 수정 또는 품목 삭제 시  서버에 저장되어 있는 이미지 삭제 구현

+++ Node.js/The Complete Guide/Section_20/controllers/admin.js
@@ -2,6 +2,7 @@
 const { validationResult } = require("express-validator");
 
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("./admin/edit-product", {
@@ -189,7 +190,10 @@ exports.postEditProduct = (req, res, next) => {
 
       product.title = updateedTitle;
       product.price = updateedPrice;
-      product.imageUrl = image?.path ?? product.imageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = updateedDesc;
 
       return product.save().then((result) => {
@@ -207,9 +211,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-
-  //Product.findByIdAndRemove(prodId)
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found."));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("Delete PRODUCT");
       res.redirect("/admin/products");
